requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
youtube-transcript-api
langchain-openai
openai
PyPDF2
unstructured
python-docx
faiss-cpu
langchain-experimental
